+++ lilith-data/logging-xml-serializer/src/test/java/de/huxhorn/lilith/data/logging/xml/LoggingXmlEventIOTest.java
@@ -1,6 +1,6 @@
 /*
  * Lilith - a log event viewer.
- * Copyright (C) 2007-2016 Joern Huxhorn
+ * Copyright (C) 2007-2017 Joern Huxhorn
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -17,7 +17,7 @@
  */
 
 /*
- * Copyright 2007-2016 Joern Huxhorn
+ * Copyright 2007-2017 Joern Huxhorn
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -44,12 +44,12 @@ import org.slf4j.LoggerFactory;
 
 import static org.junit.Assert.assertTrue;
 
-public class LoggingEventIOTest
+public class LoggingXmlEventIOTest
 	extends LoggingEventIOTestBase
 {
-	private final Logger logger = LoggerFactory.getLogger(LoggingEventIOTest.class);
+	private final Logger logger = LoggerFactory.getLogger(LoggingXmlEventIOTest.class);
 
-	public LoggingEventIOTest(Boolean logging)
+	public LoggingXmlEventIOTest(Boolean logging)
 	{
 		super(logging);
 	}