+++ lilith-data/xml-schema/schema/logging/16/index.php
@@ -1,0 +1,4 @@
+<?php
+header('Location:http://lilith.sf.net/schema/logging/16/LoggingEvent.xsd');
+?>
+