+++ log4j/converter-log4j2/src/main/java/de/huxhorn/lilith/log4j2/producer/Log4j2LoggingConverter.java
@@ -1,6 +1,6 @@
 /*
  * Lilith - a log event viewer.
- * Copyright (C) 2007-2016 Joern Huxhorn
+ * Copyright (C) 2007-2017 Joern Huxhorn
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -17,7 +17,7 @@
  */
 
 /*
- * Copyright 2007-2016 Joern Huxhorn
+ * Copyright 2007-2017 Joern Huxhorn
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -287,11 +287,7 @@ public class Log4j2LoggingConverter
 			return null;
 		}
 
-		ExtendedStackTraceElement result = new ExtendedStackTraceElement();
-		result.setClassName(ste.getClassName());
-		result.setFileName(ste.getFileName());
-		result.setLineNumber(ste.getLineNumber());
-		result.setMethodName(ste.getMethodName());
+		ExtendedStackTraceElement result = new ExtendedStackTraceElement(ste.getStackTraceElement());
 		result.setExact(ste.getExact());
 		result.setVersion(ste.getVersion());
 		result.setCodeLocation(ste.getLocation());